# target/classes/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0",
    "title" : "Book Store"
  },
  "host" : "localhost:8081",
  "basePath" : "/bookstore-back-1.0-SNAPSHOT/api",
  "tags" : [ {
    "name" : "Book"
  } ],
  "paths" : {
    "/books" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Lists all books",
        "description" : "",
        "operationId" : "getAllBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Books found"
          },
          "204" : {
            "description" : "Books not found"
          }
        }
      },
      "post" : {
        "tags" : [ "Book" ],
        "summary" : "Creates a book given a Json book representation",
        "description" : "",
        "operationId" : "createBook",
        "consumes" : [ "application/json" ],
        "responses" : {
          "210" : {
            "description" : "Book is created"
          },
          "415" : {
            "description" : "Not a Json format"
          }
        }
      }
    },
    "/books/count" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns number of all books",
        "description" : "",
        "operationId" : "countAllBooks",
        "responses" : {
          "200" : {
            "description" : "Number of books found"
          },
          "204" : {
            "description" : "No books found"
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns a book given id",
        "description" : "",
        "operationId" : "getBook",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1.0,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Book found",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Book not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Book" ],
        "summary" : "Deletes book given ID",
        "description" : "",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1.0,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Book has been deleted"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "500" : {
            "description" : "Book not found"
          }
        }
      }
    }
  },
  "definitions" : {
    "Book" : {
      "type" : "object",
      "required" : [ "isbn", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "title" : {
          "type" : "string",
          "description" : "Title of the book",
          "minLength" : 1,
          "maxLength" : 200
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the book",
          "minLength" : 1,
          "maxLength" : 10000
        },
        "unitCost" : {
          "type" : "number",
          "format" : "float",
          "description" : "Cost of one unit",
          "minimum" : 1.0
        },
        "isbn" : {
          "type" : "string",
          "description" : "ISBN code",
          "minLength" : 1,
          "maxLength" : 50
        },
        "publicationDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date of publication"
        },
        "numberOfPages" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Number of pages"
        },
        "imageURL" : {
          "type" : "string",
          "description" : "URL of image of the book"
        },
        "language" : {
          "type" : "string",
          "description" : "Language of the book",
          "enum" : [ "ENGLISH", "SWEDISH", "FRENCH", "SPANISH", "JAPANESE", "GERMAN", "CHINESE", "NORWEGIAN", "POLISH" ]
        }
      },
      "description" : "Book resource representation"
    }
  }
}